requirements.txt
python-dotenv
black
Flask
Flask-PyMongo
requests
